added new test removed old test
